test(desktop): add branch drift dialog inspector
